ajh_service_2020/7/22

+++ code/SayHi/src/main/java/com/xalz/service/impl/ActivityMemberServiceImpl.java
@@ -1,6 +1,7 @@
 package com.xalz.service.impl;
 
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,9 +9,12 @@ import org.springframework.stereotype.Service;
 
 import com.xalz.bean.Activity;
 import com.xalz.bean.ActivityMember;
+import com.xalz.bean.Message;
 import com.xalz.bean.User;
 import com.xalz.mappers.ActivityMemberMapper;
 import com.xalz.service.ActivityMemberService;
+import com.xalz.service.ActivityService;
+import com.xalz.service.MessageService;
 
 /**
  * 活动成员表操作实现类
@@ -22,6 +26,12 @@ public class ActivityMemberServiceImpl implements ActivityMemberService{
 
 	@Autowired
 	ActivityMemberMapper activityMemberMapper;
+	
+	@Autowired
+	ActivityService activityService;
+	
+	@Autowired
+	MessageService messageService;
 	/**
 	 * 获取活动的所有成员
 	 */
@@ -41,9 +51,18 @@ public class ActivityMemberServiceImpl implements ActivityMemberService{
 		if (activityMemberMapper.selectOne(activityMember) != null) {
 			return false;
 		}else {
-			if(activityMemberMapper.insert(activityMember) == 1)
+			if(activityMemberMapper.insert(activityMember) == 1) {
+//				String messInfo = "加入成功提醒: 您成功加入《";
+//				Activity activity = activityService.getActivByPrimaryKey(activId);
+//				messInfo = messInfo + activity.getActivName() + "》,开始时间为:";
+//				messInfo = messInfo + activity.getActivStarttime() + "，希望你及时到场.";
+//				Message message = new Message();
+//				message.setMessTime(new Date());
+//				message.setUserId(userId);
+//				message.setMessInfo(messInfo);
+//				messageService.createMessage(message);
 				return true;
-				else return false;
+			}else return false;
 		}
 		
 	}
@@ -52,13 +71,18 @@ public class ActivityMemberServiceImpl implements ActivityMemberService{
 	 * 删除活动成员
 	 */
 	@Override
-	public boolean deleteActvityMemberByUI(Integer userId) {
-		//通过用户编号删除活动成员
-		ActivityMember activityMember = new ActivityMember();
-		activityMember.setUserId(userId);
-		if(activityMemberMapper.delete(activityMember) != 0)
-			return true;
-			else return false;
+	public boolean deleteActvityMember(ActivityMember activityMember) {
+		//通过用户编号和活动编号删除活动成员
+		Activity activity = activityService.getActivByPrimaryKey(activityMember.getActivId());
+		if(activity.getUserId() == activityMember.getUserId()) {
+			return false;
+		}else {
+			if(activityMemberMapper.delete(activityMember) != 0)
+				return true;
+				else return false;
+		}
+		
+		
 	}
 
 	/**
